requirements.txt
flask==1.1.1
sqlalchemy==1.3.14
pytest==5.3.5
factory-boy==2.12.0